Card selection
Reorganized the card node to have a better structure for the card selection specific elements, as well as allow for a separate script to deal with unique logic. Adds signal processing to deal with what happens when a card is clicked, including turning off other card borders as well as logging the card that was chosen. At this point, nothing is done with that information, but it will be very useful as a basis for future milestones in card interactions.